删除之前先判断是否存在

+++ sync_packages.py
@@ -51,8 +51,9 @@ def sync_package(git_url: str, branch: str = None, paths: list[str] = None):
             print(f"重命名：{p_path}至：{dest_path}")
             subprocess.getoutput(f'mv {p_path} {dest_path}')
 
-    print("删除临时clone")
-    shutil.rmtree(clone_path)
+    if os.path.exists(clone_path):
+        print("删除临时clone")
+        shutil.rmtree(clone_path)
 
     print(f"同步插件{git_url}完成")
